Simplify Factory and move resolve_async (#102)
