+++ proxystore/store/utils.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 from typing import NamedTuple
 from typing import TypeVar
 
+from proxystore.proxy import is_resolved
 from proxystore.proxy import Proxy
 from proxystore.store import base
 from proxystore.store.exceptions import ProxyStoreFactoryError
@@ -35,3 +36,26 @@ def get_key(proxy: Proxy[T]) -> NamedTuple:
             f'{type(base.StoreFactory).__name__}. {type(factory).__name__} '
             'is not supported.',
         )
+
+
+def resolve_async(proxy: Proxy[T]) -> None:
+    """Begin resolving proxy asynchronously.
+
+    Useful if the user knows a proxy will be needed soon and wants to
+    resolve the proxy concurrently with other computation.
+
+    >>> ps.proxy.resolve_async(my_proxy)
+    >>> computation_without_proxy(...)
+    >>> # p is hopefully resolved
+    >>> computation_with_proxy(my_proxy, ...)
+
+    Note:
+        The asynchronous resolving functionality is implemented
+        by :class:`~proxystore.store.base.StoreFactory`. Factories that are not
+        of this type will error when used with this function.
+
+    Args:
+        proxy (Proxy): proxy instance to begin asynchronously resolving.
+    """
+    if not is_resolved(proxy):
+        proxy.__factory__.resolve_async()
